Add first html layout and css
Put flexbox layout to divide calculator into 3 row. The centre row is divided into 2 columns. This is because all the numbers
were jumbled together. They are evenly spaced by the first column doesn't stretch the length of the column. Most of the
buttons are not the same length. Added background colour and buttons colour.