Connected the 4 basic calculation functions to display result on the DOM
Added operator and numbers event listeners so that when a user inputs a number and operator, the corresponding operator function calculates.
Now the 4 operations work only when a user enters only one number (or more than one digit number) and operator and equal for result.
The add function was concatenating the 2 numbers instead of calculating it. Added a Number method to converts the inputs to numbers then
calculate.