Added else if statement to number event listener
This was added because pressing = before entering all of the numbers or an operator could cause problems. Press = before with num1 and add operator
causes infinity result.
Also, if = was pressed before num2 was entered, num2 was not becoming the value of more that 1 digit numbers. Instead it gave an NaN error.

Added num2 = result, to operator functions because if a number is clicked again, it needs to decide if it is to be a value from a previous calculation
or a new value for num2.